requirements.txt

pyqt5
python-pyqt5-hexview